date	2006.03.22.05.28.00;	author daniel;	state Exp;
git-svn-id: https://svn.isis.vanderbilt.edu/MoBIES/UDM/trunk@2625 c0adbae9-c137-0410-b742-ec4b2d2ceded

+++ src/UML GME/Interpreter/AttributeObject.cpp
@@ -363,7 +363,7 @@ void AttributeObject::addDefReal(const std::string str)
 	if (!real_int_set)
 		real_int = atol(str.c_str());
 	else
-		defval_real.push_back( (double)real_int + (double)atof(str.c_str())/ pow(10, str.size()) );
+		defval_real.push_back( (double)real_int + (double)atof(str.c_str())/ pow((double)10, (int)str.size()) );
